% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/period.R
\name{period}
\alias{period}
\title{Period of a sequence}
\usage{
period(x)
}
\arguments{
\item{x}{a vector with the generated sequence.}
}
\value{
An integer: the number of unique values in \code{x}.
}
\description{
Counts the number of distinct values in a sequence
  before it starts repeating. Useful for estimating the period of a
  pseudorandom number generator.
}
\examples{
x <- c(3, 7, 1, 3, 7, 1, 3, 7, 1)
period(x)
}
